Move installation of system dependencies to a separate script
The code to install system dependencies has now moved to system-dependencies.sh
script, each branch will have a system dependent copy of this. The script would
expose the function `install_system_dependencies` which would be used to install
the dependencies. `run.sh` which determine the current system and fetch the
correct to install the system dependencies